% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_response.R
\name{azureml_request_response}
\alias{azureml_request_response}
\title{Creates an azureml_request_response object}
\usage{
azureml_request_response(url, success, request_timestamp, request_body,
  response_timestamp, response_body)
}
\arguments{
\item{url}{character, URL of API endpoint}

\item{success}{logical, indicates if call to API succeeded}

\item{request_timestamp}{POSIXct, datetime when client sent request}

\item{request_body}{character, request body (json format)}

\item{response_timestamp}{POSIXct, datetime when client received response
or gave up}

\item{response_body}{character, response body (json format)}
}
\value{
S3 object with class \code{azureml_request_response}, with members:
\describe{
  \item{url}{character, URL of API endpoint}
  \item{success}{logical, indicates if call to API succeeded}
  \item{request_timestamp}{POSIXct, datetime when client sent request
    or gave up}
  \item{request_body}{character, request body (json format)}
  \item{response_timestamp}{datetime when client received response}
  \item{response_body}{response body (json format)}
}
}
\description{
Generally, you will create this object using the \code{\link{consume_endpoint}} function.
}
\seealso{
\code{\link{azureml_request_response.print}}
}

