% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_response.R
\name{consume_endpoint}
\alias{consume_endpoint}
\title{Remix of AzureML consume() function}
\usage{
consume_endpoint(endpoint, inputs, globalParameters = NULL, retryDelay = 10,
  .retry = 5)
}
\arguments{
\item{endpoint}{Either an AzureML web service endpoint returned by
\code{AzureML::\link[AzureML]{publishWebService}},
\code{AzureML::\link[AzureML]{endpoints}},
or simply an AzureML web service from
\code{AzureML::\link[AzureML]{services}};}

\item{inputs}{list (named), inputs to be sent as the body of the request.
This differs from (and is stricter than) the API for
\code{AzureML::\link[AzureML]{consume}}}

\item{globalParameters}{list (named), global parameters}

\item{retryDelay}{numeric, time in seconds to delay before retrying
in case of a server error}

\item{.retry}{numeric, number of tries before failing}
}
\value{
\code{\link{azureml_request_response}} object
}
\description{
This is a re-imagining of the \code{AzureML::\link[AzureML]{consume}} function.
The API and documentation is adapted from this function.
}
\seealso{
\code{\link{validate_inputs}}, \code{\link{validate_globalParameters}}
}

