% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{get_response_sample}
\alias{get_request_sample}
\alias{get_response_sample}
\title{Get sample request/response}
\usage{
get_response_sample(endpoint)

get_request_sample(endpoint)
}
\arguments{
\item{endpoint}{Either an AzureML web service endpoint returned by
\code{AzureML::\link[AzureML]{publishWebService}},
\code{AzureML::\link[AzureML]{endpoints}},
or simply an AzureML web service from
\code{AzureML::\link[AzureML]{services}};}
}
\value{
character, json format
}
\description{
For endpoints that use request/response, use this function to get
json-formatted text.
}
\seealso{
json_viewer
}

