% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_viewer.R
\name{json_viewer}
\alias{json_viewer}
\title{Create json-viewer}
\usage{
json_viewer(..., height = 250, width = "100\%")
}
\arguments{
\item{...}{arguments passed on to
\code{listviewer::\link[listviewer]{jsonedit}}}

\item{height}{character (or coercible to character),
height of the widget in pixels or percent}

\item{width}{character (or coercible to character),
width of the widget in pixels or percent}
}
\value{
HTML widget
}
\description{
This is a wrapper around the \code{listviewer::\link[listviewer]{jsonedit}}
function, used to create an interactive (HTML) explorer for json-text.
}
\seealso{
\code{listviewer::\link[listviewer]{jsonedit}}
}

