% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obscure_authorization_token}
\alias{obscure_authorization_token}
\title{Obscure an authorization token}
\usage{
obscure_authorization_token(x, replacement = "<token_goes_here>")
}
\arguments{
\item{x}{character (JSON), AzureML credentials}

\item{replacement}{character, replacement for key}
}
\value{
character (JSON), \code{x} with text replaced
}
\description{
This function takes a JSON string and replaces the `"authorization_token"`
field with some replacement text. It is useful for demonstrating credentials
without compromising security.
}
\seealso{
\code{\link{obscure_endpoint_keys}}
}

