% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obscure_endpoint_keys}
\alias{obscure_endpoint_keys}
\title{Obscure the keys in an endpoint}
\usage{
obscure_endpoint_keys(endpoint, replacement = "<key_goes_here>")
}
\arguments{
\item{endpoint}{Either an AzureML web service endpoint returned by
\code{AzureML::\link[AzureML]{publishWebService}},
\code{AzureML::\link[AzureML]{endpoints}},
or simply an AzureML web service from
\code{AzureML::\link[AzureML]{services}};}

\item{replacement}{character, replacement for key}
}
\value{
\code{endpoint}, with \code{PrimaryKey} and \code{SecondaryKey}
  obscured
}
\description{
This can be useful step to take before displaying information
about an endpoint.
}
\seealso{
\code{\link{obscure_authorization_token}}
}

