% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_response.R
\name{parse_response}
\alias{parse_response}
\title{Parses a response-body into a dataframe}
\usage{
parse_response(response_body, response_sample, element = 1)
}
\arguments{
\item{response_body}{character (JSON)}

\item{response_sample}{character (JSON)}

\item{element}{character or numeric, component of \code{response_body} result
result to return}
}
\value{
dataframe
}
\description{
Parses a response-body into a dataframe
}

