% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_response.R
\name{print.azureml_request_response}
\alias{print.azureml_request_response}
\title{Prints an azureml_request_response object}
\usage{
\method{print}{azureml_request_response}(x, ...)
}
\arguments{
\item{x}{\code{\link{azureml_request_response}} object}
}
\value{
\code{x}
}
\description{
This method returns the object passed to it, so that it can
be used in a pipe.
}

