% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_response_body.R
\name{request_body.azureml_request_response}
\alias{request_body.azureml_request_response}
\alias{response_body.azureml_request_response}
\title{Gets request/response body}
\usage{
\method{request_body}{azureml_request_response}(x, ...)

\method{response_body}{azureml_request_response}(x, ...)
}
\arguments{
\item{x}{\code{\link{azureml_request_response}} object}
}
\value{
character (JSON), body of request/response
}
\description{
Gets request/response body
}

