% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_response}
\alias{sample_request}
\alias{sample_response}
\title{Get sample request/response payloads}
\usage{
sample_response(endpoint)

sample_request(endpoint)
}
\arguments{
\item{endpoint}{\code{AzureML::\link[AzureML]{endpoints}} member or
\code{AzureML::\link[AzureML]{services}} member}
}
\value{
character, json format
}
\description{
For endpoints with request/response, these functions return sample
json files.
}

