% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{scrape_text}
\alias{scrape_text}
\title{Scrapes text from node}
\usage{
scrape_text(url, css)
}
\arguments{
\item{url}{character, URL to scrape}

\item{css}{character, css selector to use to identify node}
}
\value{
character
}
\description{
Scrapes text from node
}
\keyword{internal}

