% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_endpoint}
\alias{validate_endpoint}
\title{Validate an endpoint}
\usage{
validate_endpoint(endpoint)
}
\arguments{
\item{endpoint}{\code{AzureML::\link[AzureML]{endpoints}} member or
\code{AzureML::\link[AzureML]{services}} member}
}
\value{
\code{endpoint}
}
\description{
Will throw an error or return the \code{endpoint}. If a service is sent, its default
endpoint is returned.
}
\details{
The source code for this code
was taken from the source code AzureML package. Perhaps this function can be a part
of the AzureML package so that the source code can be "returned".
}

